style: Renamed organizations router
