+++ src/app/api/v1/__init__.py
@@ -4,4 +4,4 @@ from . import routers
 
 router = APIRouter(prefix="/v1")
 router.include_router(routers.public)
-router.include_router(routers.organization)
+router.include_router(routers.organizations)
