+++ src/app/api/v1/routers/__init__.py
@@ -1,2 +1,2 @@
-from .organization import router as organization  # noqa: F401
+from .organizations import router as organizations  # noqa: F401
 from .public import router as public  # noqa: F401
